Avances generales

+++ FootageBackend/src/main/java/com/danieloliva/FootageBackend/usuario/service/UsuarioService.java
@@ -1,12 +1,13 @@
 
 package com.danieloliva.FootageBackend.usuario.service;
 
+import com.danieloliva.FootageBackend.errores.excepciones.UserEntityException;
 import com.danieloliva.FootageBackend.service.base.BaseService;
 import com.danieloliva.FootageBackend.service.base.StorageService;
 import com.danieloliva.FootageBackend.usuario.dto.CreateUsuarioDto;
 import com.danieloliva.FootageBackend.usuario.dto.GetUsuarioDto;
-import com.danieloliva.FootageBackend.usuario.dto.UpdateUsuarioDto;
 import com.danieloliva.FootageBackend.usuario.dto.UsuarioDtoConverter;
+import com.danieloliva.FootageBackend.usuario.model.RolUsuario;
 import com.danieloliva.FootageBackend.usuario.model.Usuario;
 import com.danieloliva.FootageBackend.usuario.repository.UsuarioRepository;
 import lombok.RequiredArgsConstructor;
@@ -18,29 +19,76 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
-import java.util.UUID;
+import java.util.List;
+import java.util.Optional;
 
-@Service("userDetailsService")
+@Service("userDetailService")
 @RequiredArgsConstructor
 public class UsuarioService extends BaseService<Usuario, Long, UsuarioRepository> implements UserDetailsService {
 
+    private final PasswordEncoder passwordEncoder;
     private final UsuarioDtoConverter usuarioDtoConverter;
     private final StorageService storageService;
+    private final UsuarioRepository usuarioRepository;
 
     @Override
     public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
         return this.repositorio.findFirstByEmail(email).orElseThrow(()-> new UsernameNotFoundException(email + " no encontrado"));
     }
 
-    public GetUsuarioDto saveUser (CreateUsuarioDto createUsuarioDto) {
-        return usuarioDtoConverter.usuarioToGetUsuarioDto(usuarioDtoConverter.createUsuarioUserDtoToUsuario(createUsuarioDto));
+    public List<Usuario> findByRol (String rolUsuario) {
+        return usuarioRepository.findByRol(RolUsuario.valueOf(rolUsuario));
     }
 
-    public GetUsuarioDto saveAdmin (CreateUsuarioDto createUsuarioDto) {
-        return usuarioDtoConverter.usuarioToGetUsuarioDto(usuarioDtoConverter.createUsuarioAdminDtoToUsuario(createUsuarioDto));
+    public Usuario saveUser (CreateUsuarioDto createUsuarioDto) {
+        if (!createUsuarioDto.getPassword().equalsIgnoreCase(createUsuarioDto.getPassword2()))
+            throw new UserEntityException("Las contraseñas no coinciden");
+        else {
+
+            Usuario usuario = Usuario.builder()
+                    .nombre(createUsuarioDto.getNombre())
+                    .apellidos(createUsuarioDto.getApellidos())
+                    .username(createUsuarioDto.getUsername())
+                    .email(createUsuarioDto.getEmail())
+                    .password(passwordEncoder.encode(createUsuarioDto.getPassword()))
+                    .premium(createUsuarioDto.isPremium())
+                    .rol(RolUsuario.USER)
+                    .localizacion(createUsuarioDto.getLocalizacion())
+                    .avatar(createUsuarioDto.getAvatar())
+                    .build();
+
+            usuarioRepository.save(usuario);
+
+            return usuario;
+
+        }
     }
 
-    public Usuario editMyProfile  (UpdateUsuarioDto newUser, MultipartFile file, Usuario usuarioAuth) {
+    public Usuario saveAdmin (CreateUsuarioDto createUsuarioDto) {
+        if (!createUsuarioDto.getPassword().equalsIgnoreCase(createUsuarioDto.getPassword2()))
+            throw new UserEntityException("Las contraseñas no coinciden");
+        else {
+
+            Usuario usuario = Usuario.builder()
+                    .nombre(createUsuarioDto.getNombre())
+                    .apellidos(createUsuarioDto.getApellidos())
+                    .username(createUsuarioDto.getUsername())
+                    .email(createUsuarioDto.getEmail())
+                    .password(passwordEncoder.encode(createUsuarioDto.getPassword()))
+                    .premium(createUsuarioDto.isPremium())
+                    .rol(RolUsuario.ADMIN)
+                    .localizacion(createUsuarioDto.getLocalizacion())
+                    .avatar(createUsuarioDto.getAvatar())
+                    .build();
+
+            usuarioRepository.save(usuario);
+
+            return usuario;
+
+        }
+    }
+
+    public Usuario editProfile  (CreateUsuarioDto newUser, MultipartFile file, Usuario usuarioAuth) {
 
         usuarioAuth.setNombre(newUser.getNombre());
         usuarioAuth.setApellidos(newUser.getApellidos());
@@ -49,19 +97,27 @@ public class UsuarioService extends BaseService<Usuario, Long, UsuarioRepository
         usuarioAuth.setPremium(newUser.isPremium());
         usuarioAuth.setRol(newUser.getRol());
 
-        storageService.deleteFile(usuarioAuth.getAvatar());
+        if (!file.isEmpty()) {
+            storageService.deleteFile(usuarioAuth.getAvatar());
 
-        String filename = storageService.store(file);
+            String filename = storageService.store(file);
 
-        String uri = ServletUriComponentsBuilder.fromCurrentContextPath()
-                .path("/download/")
-                .path(filename)
-                .toUriString();
+            String uri = ServletUriComponentsBuilder.fromCurrentContextPath()
+                    .path("/download/")
+                    .path(filename)
+                    .toUriString();
+
+            usuarioAuth.setAvatar(filename);
+        }
 
         save(usuarioAuth);
 
         return usuarioAuth;
 
+    }
+
+    public List<Usuario> findAllPremium () {
+        return usuarioRepository.findByPremium();
     }
 
 }
